Add creator validation

+++ src/controllers/movieController.js
@@ -59,7 +59,11 @@ movieController.post('/:movieId/attach', async (req, res) => {
 movieController.get('/:movieId/delete', isAuth, async (req, res) => {
   const movieId = req.params.movieId;
 
+  const movie = await movieService.getOne(movieId);
   // Check if creator
+  if (!movie.creator?.equals(req.user.id)) {
+    return res.redirect('/');
+  }
 
   await movieService.delete(movieId);
   res.redirect('/');
